feat: move txns and settings to components, add ConnectWallet to account page
